refactor: address audio probe review comments
