+++ src/platforms/web/agent-browser-provider.ts
@@ -2,7 +2,10 @@ import { runCmd } from '../../utils/exec.ts';
 import { AppError } from '../../kernel/errors.ts';
 import { sleep } from '../../utils/timeouts.ts';
 import type { Rect } from '../../kernel/snapshot.ts';
-import { normalizeAudioProbeRecord } from '../../audio-probe-result.ts';
+import {
+  buildAudioProbeEvalScript,
+  normalizeAgentBrowserAudioProbeResult,
+} from './agent-browser-audio-probe.ts';
 import { normalizeAgentBrowserNetworkRequests } from './agent-browser-network.ts';
 import { normalizeAgentBrowserSnapshot } from './agent-browser-snapshot.ts';
 import {
@@ -11,13 +14,7 @@ import {
   readStringProperty,
   type JsonObject,
 } from './json-utils.ts';
-import type {
-  WebAudioProbeOptions,
-  WebAudioProbeResult,
-  WebProvider,
-  WebSnapshotOptions,
-  WebSnapshotResult,
-} from './provider.ts';
+import type { WebProvider, WebSnapshotOptions, WebSnapshotResult } from './provider.ts';
 import { mapManagedAgentBrowserError, resolveAgentBrowserTool } from './agent-browser-tool.ts';
 
 const AGENT_BROWSER = 'agent-browser';
@@ -93,293 +90,6 @@ export function createAgentBrowserWebProvider(
       );
     },
   };
-}
-
-function buildAudioProbeEvalScript(options: WebAudioProbeOptions): string {
-  return `(${audioProbeEvalScript.toString()})(${JSON.stringify({
-    action: options.action,
-    durationMs: options.durationMs,
-    bucketMs: options.bucketMs,
-    source: options.source ?? 'media-elements',
-  })})`;
-}
-
-type AudioProbePageRecord = Record<string, any>;
-
-declare const window: AudioProbePageRecord;
-declare const document: { querySelectorAll(selector: string): any[] };
-
-function audioProbeEvalScript(options: AudioProbePageRecord): unknown {
-  const key = '__agentDeviceAudioProbe';
-  const contextKey = '__agentDeviceAudioProbeContext';
-  const sourceKey = '__agentDeviceAudioProbeSources';
-  const silenceDb = -90;
-  const now = () => Date.now();
-  const dbfs = (value: number) => {
-    if (!Number.isFinite(value) || value <= 0) return silenceDb;
-    return Math.max(silenceDb, Math.min(0, Math.round(20 * Math.log10(value))));
-  };
-  const note = (probe: AudioProbePageRecord, message: string) => {
-    if (!probe.notes.includes(message)) probe.notes.push(message);
-  };
-  const scopeNote =
-    'Audio probe samples HTML media elements exposed to Web Audio; it is not whole-tab or system audio capture.';
-  const routingNote =
-    'URL-backed media elements may be routed through the probe AudioContext while they are observed.';
-  const mediaElements = () => Array.from(document.querySelectorAll('audio,video'));
-  const stopProbe = (probe: AudioProbePageRecord | undefined, reason: string) => {
-    if (!probe || probe.state === 'stopped') return probe;
-    clearInterval(probe.timer);
-    clearTimeout(probe.timeout);
-    probe.timer = undefined;
-    probe.timeout = undefined;
-    probe.state = 'stopped';
-    probe.active = false;
-    probe.reason = reason;
-    probe.stoppedAt = now();
-    for (const entry of probe.analysers) {
-      try {
-        entry.analyser.disconnect();
-        entry.source.disconnect();
-        if (entry.audible) entry.source.connect(probe.context.destination);
-      } catch {}
-    }
-    if (probe.resumeOnGesture) {
-      for (const eventName of ['click', 'pointerdown', 'keydown']) {
-        window.removeEventListener(eventName, probe.resumeOnGesture, true);
-      }
-      probe.resumeOnGesture = undefined;
-    }
-    return probe;
-  };
-  const getContext = () => {
-    const AudioContextCtor = window.AudioContext || window.webkitAudioContext;
-    if (!AudioContextCtor) return undefined;
-    const existing = window[contextKey];
-    if (existing && existing.state !== 'closed') return existing;
-    const context = new AudioContextCtor();
-    window[contextKey] = context;
-    return context;
-  };
-  const createElementAudioSource = (probe: AudioProbePageRecord, element: AudioProbePageRecord) => {
-    if (!element.currentSrc && !element.src && element.readyState === 0) return undefined;
-    try {
-      if (!window[sourceKey]) window[sourceKey] = new WeakMap();
-      let source = window[sourceKey].get(element);
-      if (!source) {
-        // createMediaElementSource permanently moves this element through the
-        // shared probe AudioContext. We keep that context open and reconnect the
-        // source to destination on stop so audible playback survives the probe.
-        source = probe.context.createMediaElementSource(element);
-        window[sourceKey].set(element, source);
-      }
-      source.disconnect();
-      return { source, audible: true };
-    } catch {
-      return undefined;
-    }
-  };
-  const createMediaStreamAudioSource = (
-    probe: AudioProbePageRecord,
-    stream: AudioProbePageRecord | undefined,
-  ) => {
-    if (!stream || typeof stream.getAudioTracks !== 'function') return undefined;
-    if (stream.getAudioTracks().length === 0) return undefined;
-    return { source: probe.context.createMediaStreamSource(stream), audible: false };
-  };
-  const createCaptureStreamAudioSource = (
-    probe: AudioProbePageRecord,
-    element: AudioProbePageRecord,
-  ) => {
-    if (typeof element.captureStream !== 'function') return undefined;
-    const stream = element.captureStream();
-    return createMediaStreamAudioSource(probe, stream);
-  };
-  const discover = (probe: AudioProbePageRecord) => {
-    const elements = mediaElements();
-    probe.mediaElementCount = elements.length;
-    for (const element of elements) {
-      if (probe.seen.has(element)) continue;
-      const sourceEntry =
-        createMediaStreamAudioSource(probe, element.srcObject) ??
-        createCaptureStreamAudioSource(probe, element) ??
-        createElementAudioSource(probe, element);
-      if (!sourceEntry) {
-        note(probe, 'Some media elements do not expose capturable audio to Web Audio.');
-        continue;
-      }
-      probe.seen.add(element);
-      const analyser = probe.context.createAnalyser();
-      analyser.fftSize = 2048;
-      sourceEntry.source.connect(analyser);
-      analyser.connect(sourceEntry.audible ? probe.context.destination : probe.sink);
-      probe.analysers.push({
-        ...sourceEntry,
-        analyser,
-        buffer: new Float32Array(analyser.fftSize),
-      });
-    }
-    probe.sourceCount = probe.analysers.length;
-    if (probe.sourceCount === 0) {
-      note(probe, 'No capturable page media audio sources were found yet.');
-    }
-  };
-  const sample = (probe: AudioProbePageRecord | undefined) => {
-    if (!probe || probe.state !== 'running') return;
-    discover(probe);
-    let totalSquares = 0;
-    let totalSamples = 0;
-    let peak = 0;
-    for (const entry of probe.analysers) {
-      entry.analyser.getFloatTimeDomainData(entry.buffer);
-      for (const value of entry.buffer) {
-        totalSquares += value * value;
-        totalSamples += 1;
-        const abs = Math.abs(value);
-        if (abs > peak) peak = abs;
-      }
-    }
-    const rms = totalSamples > 0 ? Math.sqrt(totalSquares / totalSamples) : 0;
-    const rmsDb = dbfs(rms);
-    const peakDb = dbfs(peak);
-    probe.rmsDbfs.push(rmsDb);
-    probe.peakDbfs.push(peakDb);
-    probe.heard = probe.heard || rmsDb > silenceDb || peakDb > silenceDb;
-    const maxSamples = Math.ceil(probe.durationMs / probe.bucketMs) + 2;
-    if (probe.rmsDbfs.length > maxSamples)
-      probe.rmsDbfs.splice(0, probe.rmsDbfs.length - maxSamples);
-    if (probe.peakDbfs.length > maxSamples)
-      probe.peakDbfs.splice(0, probe.peakDbfs.length - maxSamples);
-    if (now() - probe.startedAt >= probe.durationMs) stopProbe(probe, 'duration');
-  };
-  const result = (probe: AudioProbePageRecord | undefined) => {
-    const mediaCount = mediaElements().length;
-    if (!probe) return stoppedResult([scopeNote, routingNote]);
-    return {
-      audio: 'probe',
-      state: probe.state,
-      active: probe.state === 'running',
-      heard: probe.heard,
-      source: 'media-elements',
-      backend: 'agent-browser',
-      durationMs: probe.durationMs,
-      elapsedMs: Math.max(
-        0,
-        Math.min((probe.stoppedAt || now()) - probe.startedAt, probe.durationMs),
-      ),
-      bucketMs: probe.bucketMs,
-      sampleCount: probe.rmsDbfs.length,
-      mediaElementCount: mediaCount,
-      sourceCount: probe.sourceCount,
-      rmsDbfs: probe.rmsDbfs.slice(),
-      peakDbfs: probe.peakDbfs.slice(),
-      startedAt: new Date(probe.startedAt).toISOString(),
-      stoppedAt: probe.stoppedAt ? new Date(probe.stoppedAt).toISOString() : undefined,
-      reason: probe.reason,
-      notes: [scopeNote, routingNote, ...probe.notes],
-    };
-  };
-  const stoppedResult = (notes: string[]) => ({
-    audio: 'probe',
-    state: 'stopped',
-    active: false,
-    heard: false,
-    source: 'media-elements',
-    backend: 'agent-browser',
-    durationMs: Number(options.durationMs) || 10000,
-    elapsedMs: 0,
-    bucketMs: Number(options.bucketMs) || 1000,
-    sampleCount: 0,
-    mediaElementCount: mediaElements().length,
-    sourceCount: 0,
-    rmsDbfs: [],
-    peakDbfs: [],
-    notes,
-  });
-  const action = options.action || 'status';
-  let probe = window[key];
-  if (probe && probe.state === 'running' && now() - probe.startedAt >= probe.durationMs) {
-    sample(probe);
-    stopProbe(probe, 'duration');
-  }
-  if (action === 'start') {
-    if (probe) stopProbe(probe, 'restarted');
-    const context = getContext();
-    if (!context) {
-      return stoppedResult(['Web Audio API is not available in this browser context.']);
-    }
-    const sink = context.createGain();
-    sink.gain.value = 0;
-    sink.connect(context.destination);
-    probe = {
-      state: 'running',
-      active: true,
-      context,
-      sink,
-      seen: new WeakSet(),
-      analysers: [],
-      mediaElementCount: 0,
-      sourceCount: 0,
-      durationMs: Math.max(100, Number(options.durationMs) || 10000),
-      bucketMs: Math.max(100, Number(options.bucketMs) || 1000),
-      startedAt: now(),
-      stoppedAt: undefined,
-      reason: undefined,
-      heard: false,
-      rmsDbfs: [],
-      peakDbfs: [],
-      notes: [],
-    };
-    window[key] = probe;
-    try {
-      void context.resume();
-    } catch {
-      note(probe, 'AudioContext could not be resumed by the probe.');
-    }
-    probe.resumeOnGesture = () => {
-      try {
-        void context.resume();
-      } catch {
-        note(probe, 'AudioContext could not be resumed from a user gesture.');
-      }
-    };
-    for (const eventName of ['click', 'pointerdown', 'keydown']) {
-      window.addEventListener(eventName, probe.resumeOnGesture, {
-        capture: true,
-        once: true,
-      });
-    }
-    discover(probe);
-    probe.timer = setInterval(() => sample(probe), probe.bucketMs);
-    probe.timeout = setTimeout(() => stopProbe(probe, 'duration'), probe.durationMs);
-    return result(probe);
-  }
-  if (action === 'stop') {
-    if (probe) sample(probe);
-    stopProbe(probe, 'manual');
-    return result(probe);
-  }
-  if (probe) sample(probe);
-  return result(probe);
-}
-
-function normalizeAgentBrowserAudioProbeResult(data: unknown): WebAudioProbeResult {
-  const record = readAgentBrowserEvalResultRecord(data);
-  return normalizeAudioProbeRecord(record, {
-    source: 'media-elements',
-    backend: 'agent-browser',
-    durationMs: 0,
-    elapsedMs: 0,
-    bucketMs: 1000,
-    activeFallback: false,
-    mediaElementCount: 0,
-    sourceCount: 0,
-  }) as WebAudioProbeResult;
-}
-
-function readAgentBrowserEvalResultRecord(data: unknown): JsonObject {
-  if (!isJsonObject(data)) return {};
-  return isJsonObject(data.result) ? data.result : data;
 }
 
 async function runPacedScroll(
